requirements.txt
selfbotUtils~=0.0.3
pydub~=0.25.1
discord.py-self>=1.9.1